# perf(smoothcorners): do not generate new classname per image url