… src/components/Avatars/Avatar/__snapshots__/Avatar.test.tsx.snap