# Description

### Overview

기존 `smoothCorners` interpolation을 보면, prop에 backgroundImageUrl을 포함하고 있습니다. 이는, 사용처에서 서로 다른 avatarUrl 하나마다 classname 하나가 만들어진다는 것을 의미합니다. 실제로 profiler를 통해 유저챗 화면에서 fetch more를 할 때 측정해 보면, 새로운 유저의 아바타가 필요할 때마다 classname을 하나 만들고, CSSOM에 추가한 후 (`insertRules`), css가 업데이트되었으므로 전체 스타일을 재계산하는 연산이 실행됩니다. 매번 새로운 avatarUrl이 추가될 때마다 비싼 computation이 실행됩니다.

<img width="658" alt="스크린샷 2021-07-05 오후 4 22 41" src="https://user-images.githubusercontent.com/25701854/124433367-a0b18000-ddad-11eb-8221-741994d57700.png">

### Solution

이를 개선하여, avatarUrl은 styled component의 prop으로 전달하지 않도록 합니다. 대신, element에 attribute로 전달하고 `var()` 함수를 통해 읽도록 합니다. (`attr()`는 대부분의 브라우저에서 지원하지 않아 동작하지 않더라구요)

### Analysis

**측정 1**
- 측정 방법: 유저챗 전체, 열림 목록에서 idle 상태 될 때까지 기다렸다가 측정 시작, 아래로 스크롤 하면서 fetch more 5번 수행
- 측정 조건: production server, build

| AS-IS (production) | TO-BE |
| -- | -- |
| <img width="969" alt="스크린샷 2021-07-05 오후 4 05 23" src="https://user-images.githubusercontent.com/25701854/124433784-17e71400-ddae-11eb-8a91-f775d161e524.png"> | <img width="969" alt="스크린샷 2021-07-05 오후 4 06 34" src="https://user-images.githubusercontent.com/25701854/124433833-233a3f80-ddae-11eb-8b2e-df06320f117b.png"> |

| | AS-IS (production) | TO-BE | diff |
| -- | -- | -- | -- |
| Scripting | 8503 ms | 7207 ms | **-15.2%** |
| Rendering | 1855 ms | 987 ms | **-46.8%** |

**측정 2**
- 측정 방법: 팀챗 화면에서 idle 상태 될 때까지 기다렸다가 측정 시작, 유저챗 탭으로 전환
- 측정 조건: production server, build

| AS-IS (production) | TO-BE |
| -- | -- |
| <img width="969" alt="스크린샷 2021-07-05 오후 4 08 56" src="https://user-images.githubusercontent.com/25701854/124433938-4238d180-ddae-11eb-870e-4bc47063fbb5.png"> | <img width="971" alt="스크린샷 2021-07-05 오후 4 07 33" src="https://user-images.githubusercontent.com/25701854/124433977-4b29a300-ddae-11eb-86ec-ec292636da6c.png"> |

| | AS-IS (production) | TO-BE | diff |
| -- | -- | -- | -- |
| Scripting | 3845 ms | 3041 ms | **-20.9%** |
| Rendering | 942 ms | 962 ms | **+2.12%** |

### Conclusion

- 생각보다 탭 전환 시에는 드라마틱한 감소가 없었습니다. (매니저 아바타 url이 많은데 이미 팀챗 화면에서 가져와서 그런듯)
- 위의 overhead는 새로운 아바타 url이 많을수록 크기 때문에, 유저챗 fetch more를 하면서 측정을 했을 때는 예상대로 큰 실행 시간 감소가 있었습니다.

## Changes Detail
생략.

# Tests
- [ ] Jest 테스트 코드 작성 완료
- [ ] Storybook 작성 완료

## Browser Compatibility
OS / Engine 호환성을 반드시 확인해주세요.
### Windows
- [ ] Chrome - Blink
- [ ] Edge - Blink
- [ ] Firefox - Gecko (Option)
### macOS
- [x] Chrome - Blink
- [x] Edge - Blink
- [x] Safari - WebKit
- [x] Firefox - Gecko (Option)

# (Option) Issues
(연관된 PR이나 Task / 특정 시점까지 머지하면 안됨 / 번역 추가 필요 / ...)
* 이슈 없음.