@ch-builder 2021-07-07
:tada: This PR is included in version 1.0.0-next-v1.24 :tada:

The release is available on:
- [npm package (@next-v1 dist-tag)](https://www.npmjs.com/package/@channel.io/bezier-react/v/1.0.0-next-v1.24)
- [GitHub release](https://github.com/channel-io/bezier-react/releases/tag/v1.0.0-next-v1.24)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: